test(api): adapt integ test to new admin routes
